fix: Changes launch web app parent component [#1549]
